+++ src/main/java/com/example/modeladov1/service/PaisService.java
@@ -5,9 +5,12 @@ import java.util.NoSuchElementException;
 import java.util.Optional;
 
 import com.example.modeladov1.model.Categoria;
+import com.example.modeladov1.model.Municipio;
 import com.example.modeladov1.model.Pais;
 import com.example.modeladov1.repository.PaisRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -19,30 +22,47 @@ public class PaisService {
         this.repo = paisRepository;
     }
 
-    public List<Pais> getPaises() {
-        return repo.findAll();
+    public ResponseEntity<List<Pais>> getPaises() {
+        List<Pais> lista= repo.findAll();
+        if (!lista.isEmpty()) {
+            return new ResponseEntity<>(lista, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 
-    public Pais getPaisById(int id) {
-        return repo.findById(id).orElse(null);
+    public ResponseEntity<Pais> getPaisById(int id) {
+        Pais objeto = repo.findById(id).orElse(null);
+        if(objeto!=null){
+            return new ResponseEntity<>(objeto, HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 
-    public Pais savePais(Pais pais) {
-        return repo.save(pais);
+    public ResponseEntity<Pais> savePais(Pais pais) {
+        Pais objeto = repo.save(pais);
+        return new ResponseEntity<>(objeto, HttpStatus.OK);
     }
 
-    public Pais actualizarPais(Integer id_pais, Pais paisActualizada) {
+    public ResponseEntity<Pais> actualizarPais(Integer id_pais, Pais paisActualizada) {
         Optional<Pais> paisExistente = repo.findById(id_pais);
         if (paisExistente.isPresent()) {
             Pais pais = paisExistente.get();
             pais.setNombre(paisActualizada.getNombre());
-            return repo.save(pais);
+            Pais objeto = repo.save(pais);
+            return new ResponseEntity<>(objeto,HttpStatus.OK);
         } else {
-            throw new NoSuchElementException("Pais no encontrado");
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
     }
-    public void deletePais(int id) {
-        repo.deleteById(id);
+    public ResponseEntity<Pais> deletePais(int id) {
+        if(repo.findById(id).isPresent()){
+            repo.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 }
