+++ src/main/java/com/example/modeladov1/service/RolService.java
@@ -7,8 +7,11 @@ import java.util.NoSuchElementException;
 
 import com.example.modeladov1.model.Categoria;
 import com.example.modeladov1.model.Rol;
+import com.example.modeladov1.model.VentaDetalle;
 import com.example.modeladov1.repository.RolRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -16,27 +19,38 @@ public class RolService {
     @Autowired
     RolRepository repo;
 
-    public List<Rol> getAll(){
+    public ResponseEntity<List<Rol>> getAll(){
         List<Rol> roles = new ArrayList<>();
         for(Rol rol : repo.findAll()){
             roles.add(rol);
         }
-        return roles;
+        return new ResponseEntity<>(roles, HttpStatus.OK);
     }
 
-    public Rol getOne(Integer id) {
-        return repo.findById(id).orElse(null);
+    public ResponseEntity<Rol> getOne(Integer id) {
+        Rol objeto = repo.findById(id).orElse(null);
+        if(objeto!=null){
+            return new ResponseEntity<>(objeto, HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 
-    public void add(Rol rol){
-        repo.save(rol);
+    public ResponseEntity<Rol> add(Rol rol){
+        Rol objeto = repo.save(rol);
+        return new ResponseEntity<>(objeto, HttpStatus.OK);
     }
 
-    public void eliminarRol(int id) {
-        repo.deleteById(id);
+    public ResponseEntity<Rol> eliminarRol(int id) {
+        if(repo.findById(id).isPresent()){
+            repo.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 
-    public Rol actualizarRol(Integer id_rol, Rol rolActualizado) {
+    public ResponseEntity<Rol> actualizarRol(Integer id_rol, Rol rolActualizado) {
         Optional<Rol> rolExistente = repo.findById(id_rol);
 
         if (rolExistente.isPresent()) {
@@ -45,9 +59,10 @@ public class RolService {
             // Actualiza otros campos según sea necesario
 
             // Guarda el rol actualizado en la base de datos
-            return repo.save(rol);
+            Rol objeto = repo.save(rol);
+            return new ResponseEntity<>(objeto,HttpStatus.OK);
         } else {
-            throw new NoSuchElementException("Rol no encontrado");
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
     }
